requirements.txt
streamlit_drawable_canvas
streamlit_webrtc
supervision[assets]
typer
ultralytics
vidgear